Module meta_diff::core

Structs

ComputeGraph
The core of this module - this structure contains the full compute graph

ComputeNode
The main data structure of the ComputeGraph

Operator
A struct that captures all supported mathematical operations with their metadata

ParseError

Enums

BinaryOperatorType
An enum for operators which take a two parent nodes

ConstantBinaryOperatorType
An enum for operators which take two parent nodes and produce a constant

ConstantOperatorType
An enum for all operators which produce a constant

ConstantUnaryOperatorType
An enum for operators which take a single parent node and produce a constant

Dimension
Represents all possible dimensionality arguments, used in the operators

GraphError

NaryOperatorType
An enum for operators which are applied to several parent nodes

OperatorError
An error when trying to create or manipulate an operator in in consistent way

OperatorType
An enum that represents all supported mathematical operations

SpecialUnaryOperatorType
An enum for special operators which take a single parent node.

Type
Represents the five types any ComputeNode

UnaryOperatorType
An enum for operators which take a single parent node.

Constants

OPERATOR_ABS
A OperatorType = OperatorType::Unary(UnaryOperatorType::Abs)

OPERATOR_ADD
A OperatorType::Nary(NaryOperatorType::Add)

OPERATOR_CONST
A OperatorType::Constant(ConstantOperatorType::Unary(ConstantUnaryOperatorType::Const))

OPERATOR_COS
A OperatorType = OperatorType::Unary(UnaryOperatorType::Cos)

OPERATOR_COSH
A OperatorType = OperatorType::Unary(UnaryOperatorType::CosH)

OPERATOR_DIV
A OperatorType = OperatorType::Unary(UnaryOperatorType::Div)

OPERATOR_DOT
A OperatorType::Nary(NaryOperatorType::Dot)

OPERATOR_EQ
A OperatorType::Constant(ConstantOperatorType::Binary(ConstantBinaryOperatorType::Equals))

OPERATOR_EXP
A OperatorType = OperatorType::Unary(UnaryOperatorType::Exp)

OPERATOR_EYE
A OperatorType::Constant(ConstantOperatorType::Unary(ConstantUnaryOperatorType::Eye))

OPERATOR_GT
A OperatorType::Constant(ConstantOperatorType::Binary(ConstantBinaryOperatorType::GreaterThan))

OPERATOR_GTE
A OperatorType::Constant(ConstantOperatorType::Binary(ConstantBinaryOperatorType::GreaterThanOrEqual))

OPERATOR_HORZCAT
A OperatorType::Nary(NaryOperatorType::HorzCat)

OPERATOR_L1_1
A OperatorType::Unary(UnaryOperatorType::L1(Dimension::First))

OPERATOR_L1_2
A OperatorType::Unary(UnaryOperatorType::L1(Dimension::Second))

OPERATOR_L1_ALL
A OperatorType::Unary(UnaryOperatorType::L1(Dimension::All))

OPERATOR_L2_1
A OperatorType::Unary(UnaryOperatorType::L2(Dimension::First))

OPERATOR_L2_2
A OperatorType::Unary(UnaryOperatorType::L2(Dimension::Second))

OPERATOR_L2_ALL
A OperatorType::Unary(UnaryOperatorType::L2(Dimension::All))

OPERATOR_LOG
A OperatorType = OperatorType::Unary(UnaryOperatorType::Log)

OPERATOR_LT
A OperatorType::Constant(ConstantOperatorType::Binary(ConstantBinaryOperatorType::LessThan))

OPERATOR_LTE
A OperatorType::Constant(ConstantOperatorType::Binary(ConstantBinaryOperatorType::LessThanOrEqual))

OPERATOR_MAX
A OperatorType::Binary(BinaryOperatorType::Max))

OPERATOR_MDIAG
A OperatorType = OperatorType::Unary(UnaryOperatorType::MatrixDiag)

OPERATOR_MIN
A OperatorType::Binary(BinaryOperatorType::Min))

OPERATOR_MINV
A OperatorType = OperatorType::Unary(UnaryOperatorType::MatrixInverse)

OPERATOR_MUL
A OperatorType::Nary(NaryOperatorType::Mul)

OPERATOR_NEG
A OperatorType = OperatorType::Unary(UnaryOperatorType::Neg)

OPERATOR_NEQ
A OperatorType::Constant(ConstantOperatorType::Binary(ConstantBinaryOperatorType::NotEquals))

OPERATOR_NONE
A OperatorType::Constant(ConstantOperatorType::None)

OPERATOR_ONES
A OperatorType::Constant(ConstantOperatorType::Binary(ConstantBinaryOperatorType::Ones))

OPERATOR_POW
A OperatorType::Binary(BinaryOperatorType::Pow))

OPERATOR_QUAD
A OperatorType::Binary(BinaryOperatorType::Quadratic))

OPERATOR_RECT
A OperatorType = OperatorType::Unary(UnaryOperatorType::Rectifier)

OPERATOR_REPLICATEH
A OperatorType::Special(SpecialUnaryOperatorType::ReplicateHorz)

OPERATOR_REPLICATEV
A OperatorType::Special(SpecialUnaryOperatorType::ReplicateVert)

OPERATOR_RESHAPE
A OperatorType::Special(SpecialUnaryOperatorType::Reshape)

OPERATOR_SIGM
A OperatorType = OperatorType::Unary(UnaryOperatorType::Sigmoid)

OPERATOR_SIGN
A OperatorType::Constant(ConstantOperatorType::Unary(ConstantUnaryOperatorType::Sign))

OPERATOR_SIN
A OperatorType = OperatorType::Unary(UnaryOperatorType::Sin)

OPERATOR_SINH
A OperatorType = OperatorType::Unary(UnaryOperatorType::SinH)

OPERATOR_SIZE_1
A OperatorType::Constant(ConstantOperatorType::Unary(ConstantUnaryOperatorType::Size(Dimension::First)))

OPERATOR_SIZE_2
A OperatorType::Constant(ConstantOperatorType::Unary(ConstantUnaryOperatorType::Size(Dimension::Second)))

OPERATOR_SQRT
A OperatorType = OperatorType::Unary(UnaryOperatorType::Sqrt)

OPERATOR_SQUARE
A OperatorType = OperatorType::Unary(UnaryOperatorType::Square)

OPERATOR_SUBASSIGN
A OperatorType::Special(SpecialUnaryOperatorType::SubAssign)

OPERATOR_SUBINDEX
A OperatorType::Special(SpecialUnaryOperatorType::SubIndex)

OPERATOR_SUM_1
A OperatorType::Unary(UnaryOperatorType::Sum(Dimension::First))

OPERATOR_SUM_2
A OperatorType::Unary(UnaryOperatorType::Sum(Dimension::Second))

OPERATOR_SUM_ALL
A OperatorType::Unary(UnaryOperatorType::Sum(Dimension::All))

OPERATOR_TAN
A OperatorType = OperatorType::Unary(UnaryOperatorType::Tan)

OPERATOR_TANH
A OperatorType = OperatorType::Unary(UnaryOperatorType::TanH)

OPERATOR_TRANSPOSE
A OperatorType = OperatorType::Unary(UnaryOperatorType::Transpose)

OPERATOR_VDIAG
A OperatorType = OperatorType::Unary(UnaryOperatorType::VectorDiag)

OPERATOR_VERTCAT
A OperatorType::Nary(NaryOperatorType::VertCat)

OPERATOR_ZEROS
A OperatorType::Constant(ConstantOperatorType::Binary(ConstantBinaryOperatorType::Zeros))

Functions

parseMetaFile