Enum meta_diff::core::ConstantUnaryOperatorType

pub enum ConstantUnaryOperatorType {
    Const,
    Eye,
    Sign,
    Size(Dimension),
}

An enum for operators which take a single parent node and produce a constant

The operator should have a single parent and no argumetns

Variants

Const

Transforms the parent node to a constant one, regardless if whether it is dependable on any parameters or not

Eye

Creates a new identity matrix with dimensions given by the parent node

Sign

Represents elementwise sign(x)

Size

Returns the size of the node along the selected dimension. Should never be used with Dimension::All!