Enum meta_diff::core::OperatorError

pub enum OperatorError {
    ParentNotFound(OperatorType, usize, Vec<usize>),
    ArgumentNotFound(OperatorType, usize, Vec<usize>),
    AncestorNotFound(OperatorType, usize, Vec<usize>, Vec<usize>),
    InvalidNumberOfParents(OperatorType, usize, usize),
    InvaludNumberOfArguments(OperatorType, usize, usize),
    InvalidNumberOfAncestors(OperatorType, usize, usize),
    InvalidDimensionArgument(String, usize, Vec<usize>),
}

An error when trying to create or manipulate an operator in in consistent way

Variants

ParentNotFound

When trying to swap a prent which is not in the parents' list Fields are (operator, requested parent, parents' list)

ArgumentNotFound

When trying to swap an argument which is not in the parents' list Fields are (operator, requested argument, arguments' list)

AncestorNotFound

When trying to swap an ancestor which is not in the parents' list Fields are (operator, requested ancestor, parents' list, arguments' list)

InvalidNumberOfParents

When trying to create an operator with the wrong number of parents Fields are (operator, given number of parents, expected number of parents)

InvaludNumberOfArguments

When trying to create an operator with the wrong number of arguments Fields are (operator, given number of parents, expected number of parents)

InvalidNumberOfAncestors

When trying to create an operator with the wrong number of ancesotrs Fields are (operator, given number of parents, expected number of parents)

InvalidDimensionArgument

When trying to create an operator with wrongly specified dimension. These are in in integer format where 0 is Dimension::All Fields are (operator name, dimension provided, possible dimensions)