Enum meta_diff::core::SpecialUnaryOperatorType

pub enum SpecialUnaryOperatorType {
    SubIndex,
    SubAssign,
    Reshape,
    ReplicateHorz,
    ReplicateVert,
}

An enum for special operators which take a single parent node.

The operator should have one parent and several arguments

Variants

SubIndex

Takes the sub block of the parent node described by the following 4 arguments. The operator has the following syntax: (parent, start_x, sizeOfBlockX, start_y, sizeOfBlockY)

SubAssign

Represents subassignment. This will produce a matrix for which the sub block described by the 4 argument will be equal to the parent node. E.g. this means that the result is a matrix of zeros, whose subblock is equal to the parent. The operator has the following syntax: (parent, start_x, sizeOfBlockX, start_y, sizeOfBlockY)

Reshape

Represents reshaping the parent node to a size given by its arguments. The operator has the following syntax: (parent, dim_1, dim_2)

ReplicateHorz

Represents the replication of the parent node horizontally. It is assumed that it is a scalar or column vector. The operator has the following syntax: (parent, times)

ReplicateVert

Represents the replication of the parent node vertically. It is assumed that it is a scalar or row vector. The operator has the following syntax: (parent, times)

Methods

impl SpecialUnaryOperatorType

fn required_num_of_args(&self) -> usize

Returns the number of required arguments for the operator