Struct meta_diff::core::ComputeGraph

pub struct ComputeGraph {
    pub name: String,
    pub nodes: Vec<Option<ComputeNode>>,
    pub ordering: Vec<usize>,
    pub outputs: Vec<usize>,
    // some fields omitted
}

The core of this module - this structure contains the full compute graph

Fields

name

nodes

ordering

outputs

Methods

impl ComputeGraph

fn len(&self) -> usize

Returns the number of nodes in the graph

fn new() -> Self

Create a new compute graph for a function with the input name

fn add_parameter(&mut self, name: String) -> usize

Creates a new Parameter variable with the given name, inserts it in the variable table and returns its id

fn add_float(&mut self, value: f64) -> usize

Creates a new Float variable and returns its id

fn add_int(&mut self, value: i64) -> usize

Creates a new Integer variable and returns its id

fn add_const_input(&mut self, name: String) -> usize

Creates a new ConstInput variable with the given name, inserts it in the variable table and returns its id

fn add_operation(&mut self, op_type: OperatorType, args: Vec<usize>) -> Result<usize, GraphError>

Adds a variable coresponding to the input operation to the graph

fn generate_ordering(&mut self, targets: Vec<usize>) -> Result<Vec<usize>, GraphError>

Generates an ordering of computation

fn direct_gradient(&mut self) -> Result<(), GraphError>

Applies gradient operator with the first of the outputs

fn gradient(&mut self, target: usize) -> Result<(), GraphError>

Applies a gradient operator to the graph given the target

fn get_mut_node(&mut self, index: usize) -> Result<&mut ComputeNode, GraphError>

fn get_node(&mut self, index: usize) -> Result<&ComputeNode, GraphError>

fn pop_node(&mut self, index: usize) -> Result<ComputeNode, GraphError>

fn insert_node(&mut self, index: usize, node: Option<ComputeNode>) -> Option<ComputeNode>

fn is_dependable(&mut self, index: usize) -> Result<bool, GraphError>

fn swap_child_connections(&mut self, old_parent: usize, new_parent: usize) -> Result<(), GraphError>

fn string_to_operator(&mut self, name: String, args: Vec<usize>) -> Result<usize, GraphError>

fn is_function_name(name: &str) -> bool

fn get_params(&self) -> (Vec<usize>, Vec<String>)