Struct meta_diff::core::ComputeNode

pub struct ComputeNode {
    pub id: usize,
    pub node_type: Type,
    pub name: String,
    pub children: Vec<usize>,
    pub grad_level: u8,
    pub inline: bool,
    pub grad_child: Option<usize>,
    pub grad_parents: Vec<usize>,
    pub op: Operator,
}

The main data structure of the ComputeGraph

Fields

id

The id is equivalent to the index in the nodes list of the ComputeGraph

node_type

The type of the node

name

A user given or automatically generated name

children

All nodes dependable on this one

grad_level

Defines to what gradient level computation this node belongs

inline

Whether the node should be inlined by any of the source code generators

grad_child

Defines the node which represents dL/dx

grad_parents

If this node has grad_level more than 0, it means it could represent for some node dL/dx, thus this contains a list of all such nodes

op

What is the operator this node represents

Methods

impl ComputeNode

fn new(id: usize, node_type: Type, grad_level: u8, op: Operator) -> Self

Creates a new empty ComputeNode, its name depends on the input type and gradient level