Struct meta_diff::core::Operator

pub struct Operator {
    pub op_type: OperatorType,
    pub parents: Vec<usize>,
    pub args: Vec<usize>,
}

A struct that captures all supported mathematical operations with their metadata

Fields

op_type

The type of this operator

parents

List of parent nodes

args

List of argument nodes

Methods

impl Operator

fn new(op_type: OperatorType, parents: Vec<usize>, args: Vec<usize>) -> Result<Self, OperatorError>

Creates a new Operator with the given ancestors Returns an error if the given ancestors do not match the type of operator

fn get_ancestors(&self) -> Chain<Iter<usize>, Iter<usize>>

fn swap_parent(&self, old_parent: usize, new_parent: usize) -> Result<Self, OperatorError>

Swaps the given parents of the operator Returns an error if old_parent was not in the parents' list

fn swap_parent_in_place(&mut self, old_parent: usize, new_parent: usize) -> Result<(), OperatorError>

Swaps the given the parents of the operator in place Returns an error if old_parent was not in the parents' list

fn swap_argument(&self, old_arg: usize, new_arg: usize) -> Result<Self, OperatorError>

Swaps the given the arguments of the operator Returns an error if old_arg was not in the arguments' list

fn swap_argument_in_place(&mut self, old_arg: usize, new_arg: usize) -> Result<(), OperatorError>

Swaps the given the arguments of the operator in place Returns an error if old_arg was not in the arguments' list

fn swap_ancestor(&self, old_anc: usize, new_anc: usize) -> Result<Self, OperatorError>

Swaps the given the ancestors of the operator, in all occurences both in the parents' and arguments' list Returns an error if old_anc was not in the parents' or in the arguments' list

fn swap_ancestor_in_place(&mut self, old_anc: usize, new_anc: usize) -> Result<(), OperatorError>

Swaps the given the ancestors of the operator in place, in all occurences both in the parents' and arguments' list Returns an error if old_anc was not in the parents' or in the arguments' list

fn recreate(&self, parents: Vec<usize>, args: Vec<usize>) -> Result<Self, OperatorError>

Creates an operator of the same type, but with different parents and argumetns