Struct meta_diff::linking::DLFunction

pub struct DLFunction<P, R> {
    pub lib: DynamicLibrary,
    pub func: fn(P) -> R,
}

An encapsulater object of a function loaded from a dynamic library. The object is required because the if the DynamicLibrary object is dropped the function pointer will no longer be valid

Fields

lib

func

Methods

impl<P, R> DLFunction<P, R>

fn eval(&self, args: P) -> R

Evaluates the function with the given parameters