Function meta_diff::optimization::constant_folding

pub fn constant_folding(graph: &mut ComputeGraph) -> Result<bool, GraphError>

Performs constant folding on all nodes of the graph

This means that all nodes which are of types Int(x) or Float(x) are combined where possible. Also nodes of the from OPERATOR_CONST(x) where x is by itself ConstInput or ConstDerivde are unfolded

Returns true if the graph has been modified, false otherwise. An error if any of the operations brings up a GraphError